feat: expose types for BeeDebug
