+++ src/modules/debug/connectivity.ts
@@ -1,12 +1,5 @@
 import { safeAxios } from '../../utils/safeAxios'
-
-export interface NodeAddresses {
-  overlay: string
-  underlay: string[]
-  ethereum: string
-  public_key: string
-  pss_public_key: string
-}
+import { NodeAddresses, Peer } from '../../types'
 
 export async function getNodeAddresses(url: string): Promise<NodeAddresses> {
   const response = await safeAxios<NodeAddresses>({
@@ -16,19 +9,15 @@ export async function getNodeAddresses(url: string): Promise<NodeAddresses> {
 
   return response.data
 }
-
-export interface Peer {
-  address: string
-}
-export interface Peers {
+interface Peers {
   peers: Peer[]
 }
 
-export async function getPeers(url: string): Promise<Peers> {
+export async function getPeers(url: string): Promise<Peer[]> {
   const response = await safeAxios<Peers>({
     url: url + '/peers',
     responseType: 'json',
   })
 
-  return response.data
+  return response.data.peers
 }
